refactor: rename .abc to .core and mark elder duplciated
