--- docs/source/api_abc.rst
@@ -1,11 +1,0 @@
-.. automodule:: data_extractor.abc
-
-.. autoclass:: data_extractor.abc.SimpleExtractorMeta
-
-.. autoclass:: data_extractor.abc.ComplexExtractorMeta
-
-.. autoclass:: data_extractor.abc.AbstractSimpleExtractor
-    :members:
-
-.. autoclass:: data_extractor.abc.AbstractComplexExtractor
-    :members:
