+++ docs/source/api_core.rst
@@ -1,0 +1,11 @@
+.. automodule:: data_extractor.core
+
+.. autoclass:: data_extractor.core.SimpleExtractorMeta
+
+.. autoclass:: data_extractor.core.ComplexExtractorMeta
+
+.. autoclass:: data_extractor.core.AbstractSimpleExtractor
+    :members:
+
+.. autoclass:: data_extractor.core.AbstractComplexExtractor
+    :members:
